Fixed directories. Added CustomScrollView to HomePage.
